+++ R/coalesce.R
@@ -13,27 +13,16 @@
 #' @examples
 #' # Use a single value to replace all missing values
 #' x <- sample(c(1:5, NA, NA, NA))
-#' coalesce(x, 0)
+#' coalesce(x, 0L)
 #'
 #' # Or match together a complete vector from missing pieces
 #' y <- c(1, 2, NA, NA, 5)
 #' z <- c(NA, NA, 3, 4, 5)
 #' coalesce(y, z)
 coalesce <- function(x, ...) {
-  n <- length(x)
-
   values <- list(...)
   for (i in seq_along(values)) {
-    val <- values[[i]]
-    val_n <- length(val)
-
-    if (val_n == 1L) {
-      x[is.na(x)] <- val
-    } else if (val_n == n) {
-      x[is.na(x)] <- val[is.na(x)]
-    } else {
-      stop("Vector at position ", i, " is not length 1 or ", n, call. = FALSE)
-    }
+    x <- replace_with(x, is.na(x), values[[i]], paste0("Vector ", i))
   }
   x
 }